+++ app/src/main/java/com/deslomator/tagtimer/viewmodel/LabelPreselectionViewModel.kt
@@ -25,7 +25,7 @@ class LabelPreselectionViewModel @Inject constructor(
     private val appDao: AppDao, savedStateHandle: SavedStateHandle
 ): ViewModel() {
 
-    private val _sessionId = MutableStateFlow(0)
+    private val _sessionId = MutableStateFlow(0L)
     private val _state = MutableStateFlow(LabelPreselectionState())
 
     private val _tags = appDao.getActiveTags()
@@ -251,7 +251,7 @@ class LabelPreselectionViewModel @Inject constructor(
     }
 
     init {
-        val sessionId = savedStateHandle.get<Int>("sessionId") ?: 0
+        val sessionId = savedStateHandle.get<Long>("sessionId") ?: 0
         _sessionId.update { sessionId }
         viewModelScope.launch {
             _state.update {
